requirements.txt
python-dotenv==0.15.0
Flask==1.1.2
Flask-PyMongo==2.3.0
pytest==6.2.2